\name{data20}
\docType{data}
\alias{data20}
\title{Artificial nominal dataset}
\description{
  This dataset consists of 5 nominal variables and 20 cases.
  Its main aim is to demonstrate the desired entry data structure for the
  nomclust package.
  }
\usage{data(data20)}
\format{A data frame containing 5 variables and 20 cases.}
\source{created by the authors of the nomclust package}
\keyword{datasets}