% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm.R
\name{vm}
\alias{vm}
\title{Variable Mutability (VM) measure}
\usage{
vm(data)
}
\arguments{
\item{data}{A data.frame or a matrix with cases in rows and variables in colums.}
}
\value{
The function returns an object of the class "dist".
\cr
}
\description{
The function calculates a dissimilarity matrix based on the VM similarity measure.
\cr
}
\details{
The Variable Mutability similarity measure was introduced in (Sulc and Rezankova, 2019).
It treats the similarity between two categories based on the within-cluster variability expressed by the normalized mutability. The measure assigns higher weights to rarer categories.
}
\examples{
#sample data
data(data20)

# dissimilarity matrix calculation
prox.vm <- vm(data20)

}
\references{
Sulc Z. and Rezankova H. (2019). Comparison of Similarity Measures for Categorical Data in Hierarchical Clustering. Journal of Classification. 2019, 35(1), p. 58-72. DOI: 10.1007/s00357-019-09317-5.
}
\seealso{
\code{\link[nomclust]{eskin}},
\code{\link[nomclust]{good1}},
\code{\link[nomclust]{good2}},
\code{\link[nomclust]{good3}},
\code{\link[nomclust]{good4}},
\code{\link[nomclust]{iof}},
\code{\link[nomclust]{lin}},
\code{\link[nomclust]{lin1}},
\code{\link[nomclust]{of}},
\code{\link[nomclust]{sm}},
\code{\link[nomclust]{ve}}.
}
\author{
Zdenek Sulc. \cr Contact: \email{zdenek.sulc@vse.cz}
}
